Listo cancelar personalRecord cuando no se quiere crear

+++ Acme-HandyWorker/src/main/java/services/CurriculumService.java
@@ -102,8 +102,8 @@ public class CurriculumService {
 		final Collection<EndorserRecord> endorserRecords = new ArrayList<>(this.endorserRecordService.findEndorserRecordByCurriculumId(curriculum.getId()));
 		final Collection<MiscellaneousRecord> miscellaneousRecords = new ArrayList<>(this.miscellaneousRecordService.findMiscellaneousRecordByCurriculumId(curriculum.getId()));
 
-		this.personalRecordService.delete(personalRecord);
-		this.tickerService.delete(curriculum.getTicker());
+		if (personalRecord != null)
+			this.personalRecordService.delete(personalRecord);
 
 		if (educationRecords.size() != 0 || professionalRecords.size() != 0 || endorserRecords.size() != 0 || miscellaneousRecords.size() != 0) {
 			for (final EducationRecord educationRecord : educationRecords)
